soo, it was like fighting against Lernaean Hydra...
now, jamocha compiles. but nothing more. starting it will lead to an exception. now there is a lot of bugfixing needed for getting it work again.

- some more old nodes deleted
- basicrulecompiler deleted (since it only would give a lot of trouble to adapt it to our new stuff)
- in workingmemory[impl]? ive commented out some functions, which arent really needed now, since our new nodes will handle memory in a different manner

IMPORTANT: for now, jamocha compiles. thats great. but dont expect that something really works for now! it wouldnt! the whole caboodle has to become adapted the next days...

git-svn-id: https://jamocha.svn.sourceforge.net/svnroot/jamocha/jamocha@974 2cd91b7d-3b21-0410-a5f4-f42db4f24360