i've tried to implement binding generation...
imo, the general approach is not bad and you can observe, that the actual implementation something does things right; but for now, its not wrong to say "it doesnt work".

additionally, ive to confess that the new code is dirty hacked (typecasting-hell and so on) and not really clean, since it was only a proof-of-concept for me.

but for today, i'm a litte depressed ^^ and so i suspend work till tomorrow...

git-svn-id: https://jamocha.svn.sourceforge.net/svnroot/jamocha/jamocha@1047 2cd91b7d-3b21-0410-a5f4-f42db4f24360