written a little framework for our testing:
in my last commit, i added the "bierwurstsalat" example folder.

now, i've created an ant task, which executes the test and compares it with the expected results (which also are included in rules.clp as a special type of comment).

so, now we are able to write test rules which automatically can be compared to expected results. for really making use of this, we have to fill examples/bierwurstsalat/rules.clp with maaany test rules and their expected results. like already written, dont worry to fill in some tests here, too!

important: please fulfill my conventions in the file:
	* "[rulename]" at the beginning of all printouts.
	* comment out not working rules by single ";"
	* write expected result below the rule by double ";;"
	* verify the expected results when making copy&paste, so we can assume, these results are really the correct one!
	* in most cases, each test rule only should make a printout. any additional funny magic things on the right hand side aren't able to become verified.
	
have fun ;)

git-svn-id: https://jamocha.svn.sourceforge.net/svnroot/jamocha/jamocha@1376 2cd91b7d-3b21-0410-a5f4-f42db4f24360